tools/texture-editor/requirements.txt
PyQt6>=6.6.0
PyOpenGL>=3.1.7
Pillow>=10.0.0
PyYAML>=6.0
numpy>=1.21.0
